+++ login-workflow/src/index.ts
@@ -12,3 +12,4 @@ export { default as i18n } from './translations/i18n';
 export * from '@brightlayer-ui/react-auth-shared';
 
 export * from './screens';
+export * from './components';