+++ login-workflow/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v3.2.0 (Unreleased)
+
+### Added
+
+-   A `BrandedCardContainer` component.
+-   Confirm password field error state message in `ChangePasswordForm`.
+
+### Changed
+
+-   Changed the verification code styles for the self registration workflow.
+
 ## v3.1.1 (December 14, 2022)
 
 ### Changed